+++ classes/fields/font/in_dashboard.tpl.php
@@ -28,7 +28,7 @@
         ?>
         <label for="tea_add_contents_<?php echo $number ?>_font_<?php echo $ft[0] ?>" class="gfont_<?php echo str_replace(' ', '_', $ft[1]) ?> <?php echo $selected ? 'selected' : '' ?>">
             <span>
-                <input type="radio" name="tea_add_contents[<?php echo $number ?>][std]" id="tea_add_contents_<?php echo $number ?>_font_<?php echo $ft[0] ?>" value="<?php echo $ft[0] ?>" <?php echo $selected ? 'checked="checked" ' : '' ?> />
+                <input type="radio" name="tea_add_contents[<?php echo $number ?>][stands]" id="tea_add_contents_<?php echo $number ?>_font_<?php echo $ft[0] ?>" value="<?php echo $ft[0] ?>" <?php echo $selected ? 'checked="checked" ' : '' ?> />
                 <b><?php echo $ft[0] ?></b>
             </span>
         </label>
@@ -40,7 +40,37 @@
     <span><?php _e('Custom Google Font', TTO_I18N) ?></span>
 </label>
 
-<div class="label-edit-options" style="display:block;">
+<div class="label-edit-options model-multiple" style="display:block;">
+    <label class="label-option">
+        <?php _e('Default', TTO_I18N) ?>
+    </label>
+    <label class="label-third">
+        <?php _e('Google name', TTO_I18N) ?>
+    </label>
+    <label class="label-third">
+        <?php _e('CSS name', TTO_I18N) ?>
+    </label>
+    <label class="label-third">
+        <?php _e('Size(s)', TTO_I18N) ?>
+    </label>
+    <div class="clearfix"></div>
+
+    <div class="label-model">
+        <label class="label-option">
+            <input type="radio" name="tea_add_contents[<?php echo $number ?>][stands]" value="__OPTNUM__" />
+        </label>
+        <label class="label-third">
+            <input type="text" name="tea_add_contents[<?php echo $number ?>][options][__OPTNUM__][]" value="" class="code" placeholder="<?php _e('Google name, ex: &quot;PT+Sans&quot;', TTO_I18N) ?>" />
+        </label>
+        <label class="label-third">
+            <input type="text" name="tea_add_contents[<?php echo $number ?>][options][__OPTNUM__][]" value="" class="code" placeholder="<?php _e('CSS Name, ex: &quot;PT Sans&quot;', TTO_I18N) ?>" />
+        </label>
+        <label class="label-third">
+            <input type="text" name="tea_add_contents[<?php echo $number ?>][options][__OPTNUM__][]" value="" class="code" placeholder="<?php _e('Size(s), ex: &quot;400,700&quot;', TTO_I18N) ?>" />
+        </label>
+        <div class="clearfix"></div>
+    </div>
+
     <?php
         $optnum = 0;
 
@@ -53,6 +83,9 @@
                     continue;
                 }
         ?>
+            <label class="label-option">
+                <input type="radio" name="tea_add_contents[<?php echo $number ?>][stands]" value="<?php echo $optnum+1 ?>" <?php echo $std == $opts[0] ? 'checked="checked"' : '' ?> />
+            </label>
             <label class="label-third">
                 <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $k ?>][]" value="<?php echo isset($opts[0]) ? $opts[0] : '' ?>" class="code" placeholder="<?php _e('Google name, ex: &quot;PT+Sans&quot;', TTO_I18N) ?>" />
             </label>
@@ -66,14 +99,6 @@
             <?php $optnum++; ?>
         <?php endforeach ?>
     <?php endif ?>
-    <label class="label-third">
-        <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $optnum ?>][]" value="" class="code" placeholder="<?php _e('Google name, ex: &quot;PT+Sans&quot;', TTO_I18N) ?>" />
-    </label>
-    <label class="label-third">
-        <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $optnum ?>][]" value="" class="code" placeholder="<?php _e('CSS Name, ex: &quot;PT Sans&quot;', TTO_I18N) ?>" />
-    </label>
-    <label class="label-third">
-        <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $optnum ?>][]" value="" class="code" placeholder="<?php _e('Size(s), ex: &quot;400,700&quot;', TTO_I18N) ?>" />
-    </label>
-    <div class="clearfix"></div>
+
+    <input type="submit" name="add_option" class="button-secondary label-button" value="<?php _e('Add a new option', TTO_I18N) ?>" />
 </div>