+++ classes/fields/select/in_dashboard.tpl.php
@@ -13,15 +13,29 @@
 </label>
 
 <label class="label-edit-content">
-    <span><?php _e('Default value', TTO_I18N) ?></span>
-    <input type="text" name="tea_add_contents[<?php echo $number ?>][std]" value="<?php echo $std ?>" class="code" />
+    <b><?php _e('Options', TTO_I18N) ?></b>
 </label>
 
-<label class="label-edit-content">
-    <span><?php _e('Options', TTO_I18N) ?></span>
-</label>
+<div class="label-edit-options model-multiple" style="display:block;">
+    <label class="label-option">
+        <?php _e('Default', TTO_I18N) ?>
+    </label>
+    <label class="label-first">
+        <?php _e('Label', TTO_I18N) ?>
+    </label>
+    <div class="clearfix"></div>
 
-<div class="label-edit-options" style="display:block;">
+    <div class="label-model">
+        <label class="label-option">
+            <input type="radio" name="tea_add_contents[<?php echo $number ?>][default]" value="__OPTNUM__" />
+        </label>
+        <label class="label-first">
+            <input type="hidden" name="tea_add_contents[<?php echo $number ?>][options][__OPTNUM__][]" value="" />
+            <input type="text" name="tea_add_contents[<?php echo $number ?>][options][__OPTNUM__][]" value="" class="code" placeholder="<?php _e('Your label option', TTO_I18N) ?>" />
+        </label>
+        <div class="clearfix"></div>
+    </div>
+
     <?php
         $optnum = 0;
 
@@ -37,21 +51,17 @@
                 $vallabel = !empty($opts[1]) ? $opts[1] : $opts[0];
                 $valvalue = $opts[0];
         ?>
-            <label class="label-second">
-                <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $k ?>][]" value="<?php echo $valvalue ?>" class="code" placeholder="<?php _e('Your value option', TTO_I18N) ?>" />
+            <label class="label-option">
+                <input type="radio" name="tea_add_contents[<?php echo $number ?>][default]" value="<?php echo $optnum+1 ?>" <?php echo $std == $valvalue ? 'checked="checked"' : '' ?> />
             </label>
-            <label class="label-second">
+            <label class="label-first">
+                <input type="hidden" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $k ?>][]" value="<?php echo $valvalue ?>" />
                 <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $k ?>][]" value="<?php echo $vallabel ?>" class="code" placeholder="<?php _e('Your label option', TTO_I18N) ?>" />
             </label>
             <div class="clearfix"></div>
             <?php $optnum++; ?>
         <?php endforeach ?>
     <?php endif ?>
-    <label class="label-second">
-        <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $optnum ?>][]" value="" class="code" placeholder="<?php _e('Your value option', TTO_I18N) ?>" />
-    </label>
-    <label class="label-second">
-        <input type="text" name="tea_add_contents[<?php echo $number ?>][options][<?php echo $optnum ?>][]" value="" class="code" placeholder="<?php _e('Your label option', TTO_I18N) ?>" />
-    </label>
-    <div class="clearfix"></div>
+
+    <input type="submit" name="add_option" class="button-secondary label-button" value="<?php _e('Add a new option', TTO_I18N) ?>" />
 </div>