+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.5 (August 23, 2013)
+
+Better integration in fields
+
+
 ## 1.4.4 (August 22, 2013)
 
 Add new uninstall.php to make uninstall works